backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
pydantic-settings
python-multipart
python-jose[cryptography]
passlib[bcrypt]
openai
scikit-learn
pandas
mediapipe
opencv-python-headless
numpy
python-dotenv
cors
